feat: Add promotional banners
Introduces a new `PromoBanner` type and associated constants and services to manage promotional banners. This change lays the groundwork for displaying dynamic promotions on the homepage and potentially other areas of the application.

Additionally, this commit refactors the `Navbar` component to use "Usuario" instead of "Admin" for the user profile link and adjusts layout for better responsiveness.